+++ app/Http/Controllers/User/UserCommentController.php
@@ -15,6 +15,7 @@ use App\Support\Filter;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
+use Illuminate\Support\Facades\Gate;
 use Symfony\Component\HttpFoundation\Response as ResponseCodes;
 
 class UserCommentController extends Controller
@@ -47,12 +48,19 @@ class UserCommentController extends Controller
      * Store a newly created resource in storage.
      *
      * @param StoreProductCommentRequest $request
+     * @param Product $product
      * @return JsonResponse
      */
     public function store(StoreProductCommentRequest $request, Product $product): JsonResponse
     {
-        $validated = $request->validated(['pros', 'cons', 'description']);
-        $model = $this->service->create(['product' => $product->id] + $validated);
+        Gate::authorize('create', [Comment::class, $product]);
+
+        $validated = filter_validated_data($request->validated(), [
+            'pros',
+            'cons',
+            'description',
+        ]);
+        $model = $this->service->create($validated + ['product' => $product->id]);
 
         if (!is_null($model)) {
             return response()->json([
@@ -75,6 +83,7 @@ class UserCommentController extends Controller
      */
     public function show(Comment $comment): UserProductCommentSingleResource
     {
+        Gate::authorize('canDoOperation', $comment);
         return new UserProductCommentSingleResource($comment);
     }
 
@@ -90,7 +99,14 @@ class UserCommentController extends Controller
         Comment                         $comment
     ): UserProductCommentSingleResource|JsonResponse
     {
-        $validated = $request->validated(['product', 'pros', 'const', 'description']);
+        Gate::authorize('canDoOperation', [$comment, true]);
+
+        $validated = filter_validated_data($request->validated(), [
+            'product',
+            'pros',
+            'const',
+            'description',
+        ]);
         $model = $this->service->updateById($comment->id, $validated);
 
         if (!is_null($model)) {
@@ -111,6 +127,8 @@ class UserCommentController extends Controller
      */
     public function destroy(Request $request, Comment $comment): JsonResponse
     {
+        Gate::authorize('canDoOperation', $comment);
+
         $res = $this->service->deleteUserCommentById($request->user()->id, $comment->id);
 
         if ($res) {
